Extensions added

+++ AntSmasher/app/src/main/java/com/ananddamodaran/antsmasher/Ant.kt
@@ -3,4 +3,4 @@ package com.ananddamodaran.antsmasher
 /**
  * Created by anand on 3/16/18.
  */
-class Ant(val id:Long, val x: Float , val y: Float)
+class Ant(val id:Int, val x: Double , val y: Double)
